+++ CHANGELOG.rst
@@ -34,6 +34,7 @@ UC
 
     * data pullers
     * auto_off unit state processor can be cancelled / modified while running
+    * "server cleanup" CLI command (cleans state db)
 
 LM PLC
 ------